refactor : 스타일 컴포넌트 및 적용된 컴포넌트 이름 수정, 적용
