+++ src/Component/Header/Modal/CalendarDetail/Calendar/Calendar.tsx
@@ -1,7 +1,7 @@
 import { PrevButton, NextButton } from "util/Icons";
-import { StyledCalendar } from "../CalendarDetail.styled";
-import CalTitle from "./CalTitle";
-import CalDayBox from "./CalDayBox";
+import { StyledCalendar } from "./Calendar.styled";
+import CalendarTitle from "./CalendarTitle";
+import CalendarDayBox from "./CalendarDayBox";
 
 const Calendar = ({
 	title,
@@ -19,11 +19,11 @@ const Calendar = ({
 			{title === "prev" && (
 				<PrevButton onClick={() => changeDates(title)} colorset="black" size={16} />
 			)}
-			<CalTitle year={year} month={month} />
+			<CalendarTitle year={year} month={month} />
 			{title === "next" && (
 				<NextButton onClick={() => changeDates(title)} colorset="black" size={16} />
 			)}
-			<CalDayBox year={year} month={month} />
+			<CalendarDayBox year={year} month={month} />
 		</StyledCalendar>
 	);
 };
